% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rule.R
\name{add_rule}
\alias{add_rule}
\title{Add a Rule to a JPG Picture}
\usage{
add_rule(image, height, min = 0, max = 140, path)
}
\arguments{
\item{image}{(string) Path to the picture}

\item{height}{(numeric) Real height of the feature}

\item{min}{(numeric) Minimum value of the rule}

\item{max}{(numeric) Maximum value of the rule}

\item{path}{(string) Folder to store picture}
}
\description{
This function adds a rule to a JPG picture and a marker of the feature height
}
\examples{
##
}
